requirements.txt
flask
flask-sqlalchemy
flask-cors
python-dotenv
authlib
requests
